// web/src/api/generated/model/moduleSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ModuleNextClass } from './moduleNextClass';

export interface ModuleSummary {
  code: string;
  name: string;
  lecturer: string | null;
  next_class: ModuleNextClass | null;
  new_notes: number;
}